OC-1045 - Publication edit page
