+++ ui/src/components/index.tsx
@@ -75,6 +75,7 @@ export { default as PublicationCreationLinkedItems } from './Publication/Creatio
 export { default as PublicationCreationMainText } from './Publication/Creation/MainText';
 export { default as PublicationCreationResearchProcess } from './Publication/Creation/ResearchProcess';
 export { default as PublicationCreationStepTitle } from './Publication/Creation/StepTitle';
+export { default as PublicationBundleForm } from './PublicationBundle/PublicationBundleForm';
 export { default as PublicationLink } from './Publication/Link';
 export { default as PublicationPageCoAuthoringActions } from './Publication/PublicationPage/CoAuthoringActions';
 export { default as PublicationPageHeaderActions } from './Publication/PublicationPage/HeaderActions';
